carray: properly set capacity of dynamic array
This patch fixes a nasty bug which corrupted the heap due to
improperly setting the array's capacity to the number of bytes to be
allocated, instead of just the number of elements. The fix is simple:
set the capacity to the correct value, and not the total number of bytes
to be allocated.